extract manifest handling to own file
