+++ internal/connection/octant.go
@@ -7,44 +7,11 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/mydecisive/mdai-gateway/internal/integration"
-	"net/http"
-	"text/template"
-
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
-	"sigs.k8s.io/yaml"
+	"net/http"
 )
-
-//go:embed templates/argo-app.yaml
-var argoAppTemplate string
-
-//go:embed templates/primary-collector.yaml
-var primaryCollectorTemplate string
-
-//go:embed templates/shadow-collector.yaml
-var shadowCollectorTemplate string
-
-//go:embed templates/envoy-config.yaml
-var envoyConfigTemplate string
-
-//go:embed templates/envoy-deployment.yaml
-var envoyDeploymentTemplate string
-
-//go:embed templates/envoy-service.yaml
-var envoyServiceTemplate string
-
-//go:embed templates/secret.yaml
-var secretTemplate string
-
-var manifestTemplates = map[string]string{
-	"primary-collector": primaryCollectorTemplate,
-	"shadow-collector":  shadowCollectorTemplate,
-	"envoy-config":      envoyConfigTemplate,
-	"envoy-deployment":  envoyDeploymentTemplate,
-	"envoy-service":     envoyServiceTemplate,
-	"secret":            secretTemplate,
-}
 
 // FIXME: Actually wire up all needed fields
 type OctantConnectionData struct {
@@ -84,15 +51,6 @@ func NewOctantConnection(httpClient *http.Client, k8sClient kubernetes.Interface
 		},
 	}
 }
-
-type ArgoTemplateData struct {
-	AppName        string
-	Namespace      string
-	ConnectionData OctantConnectionData
-	TempDDAPIKey   string
-	TempDDURL      string
-}
-
 func (oc *OctantConnection) GetConnectionByName(ctx context.Context, namespace, name string) (*OctantConnectionData, error) {
 	configmap, err := oc.K8sClient.CoreV1().ConfigMaps(namespace).Get(ctx, connectionsConfigmapName, metav1.GetOptions{})
 	if err != nil {
@@ -269,47 +227,4 @@ func (oc *OctantConnection) doArgoAppCreation(ctx context.Context, templateData 
 func (oc *OctantConnection) deleteArgoApp(ctx context.Context, namespace, name string) error {
 	// DELETE APP
 	return nil
-}
-
-func (oc *OctantConnection) renderArgoAppManifest(templateData *ArgoTemplateData) ([]byte, error) {
-	appManifestTemplate, err := template.New("argo-app").Parse(argoAppTemplate)
-	if err != nil {
-		return []byte{}, err
-	}
-	var renderedYaml bytes.Buffer
-	if err := appManifestTemplate.Execute(&renderedYaml, templateData); err != nil {
-		return []byte{}, err
-	}
-
-	renderedJson, err := yaml.YAMLToJSON(renderedYaml.Bytes())
-	if err != nil {
-		return []byte{}, err
-	}
-
-	return renderedJson, nil
-}
-
-func (oc *OctantConnection) renderSyncManifests(templateData *ArgoTemplateData) ([]string, error) {
-	// FIXME: Actually wire up telemetry types; for now just renders all three. Probably need to do more than a string template for the collector configs
-
-	var manifests []string
-	for templateName, templateString := range manifestTemplates {
-		appManifestTemplate, err := template.New(templateName).Parse(templateString)
-		if err != nil {
-			return manifests, err
-		}
-		var renderedYaml bytes.Buffer
-		if err := appManifestTemplate.Execute(&renderedYaml, templateData); err != nil {
-			return manifests, err
-		}
-
-		renderedJson, err := yaml.YAMLToJSON(renderedYaml.Bytes())
-		if err != nil {
-			return manifests, err
-		}
-
-		manifests = append(manifests, string(renderedJson))
-	}
-
-	return manifests, nil
 }
